Refactoring NotificationsProvider
